requirements/default.txt
setuptools
south==0.8.2
Django>=1.9.4
django-configurations>=2.0
dj-database-url>=0.3.0
pylibmc>=1.3.0
django-storages>=1.1.8
Pillow>=3.0.0
django-cache-url>=0.8.0
werkzeug>=0.9.4
gunicorn>=0.17.4
easy-thumbnails>=1.2
django-debug-toolbar>=1.1
django-extensions>=1.3.4
django-braces>=1.4.0
django-allauth>=0.16.1
django-floppyforms>=1.1.1
django-custom-user>=0.4
django-markdown-deux>=1.0.4
django-avatar>=2.0
django-localflavor>=1.0.0
gittle>=0.4.0
celery[redis]>=3.1.16,<4
django-celery>=3.1.16
boto>=2.9.5
django-bower>=5.0.1
django-bootstrap3>=4.8.1
raven>=5.0.0
pytz>=2014.7
shortuuid>=0.4.2
